Rename (and migrate) MenuBarInput -> MenuListButton
