+++ frontend/src/components/Editor.svelte
@@ -275,7 +275,7 @@
 	.working-colors-button .swatch > button,
 	.radio-input button,
 	.menu-list,
-	.menu-bar-input .entry,
+	.menu-list-button .entry,
 	.layer-tree .expand-arrow,
 	.widget-section .header {
 		&:focus-visible {
